requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.5
djangorestframework==3.14.0
pip==22.3.1
pycodestyle==2.10.0
pytz==2022.7
setuptools==65.5.0
sqlparse==0.4.3